First set of files - origionals
These are the first set. We will upload more to complete the project.
 These files are working as is. and have been tested on Android and desktop systems.

the purpose of the repo is to expand the A2HS function to Apple devices who are currently not supporting this feature.   So we will be developing out that feature for our project, but then exporting the feature ( once tested ) as a package for distrib to other developers who may need this to work as well.